requirements.txt
numpy
gradio
pywaggle[all]
